data/ServiceAPI/requirements.txt
anyio==3.6.2
fastapi==0.85.1
uvicorn[standard]
idna==3.4
numpy==1.23.4
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.74
ffmpeg-python==0.2.0
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
Pillow
requests